content/html/document/test/test_bug765780.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=765780
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 765780</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">
    /** Test for Bug 765780 **/
    SimpleTest.waitForExplicitFinish();
    window.onload = function() {
      var f = $("f");
      var doc = f.contentDocument;
      doc.designMode = "on";
      var s = doc.createElement("script");
      s.textContent = "parent.called = true;";

      window.called = false;
      doc.body.appendChild(s);
      ok(called, "Script in designMode iframe should have run");

      doc = doc.querySelector("iframe").contentDocument;
      var s = doc.createElement("script");
      s.textContent = "parent.parent.called = true;";

      window.called = false;
      doc.body.appendChild(s);
      ok(called, "Script in designMode iframe's child should have run");

      SimpleTest.finish();
    }
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=765780">Mozilla Bug 765780</a>
<!-- Important: iframe needs to not be display: none -->
<p id="display"><iframe id="f" src="data:text/html,<iframe></iframe>"></iframe> </p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>
</html>
